Scheduling service (#13)
This PR adds a scheduling service. It has one function that takes a set of matches and a
number of available tables and returns a schedule where each match is placed into
a time slot. A time slot contains one or more matches.

The schedule takes into account so we do not schedule more matches in one time
slot than there are available tables. And also so no player is assigned to the same
time slot more than once (i.e. getting double booked)