Cleaning up DTOs (#43)
Renamed PlayerEntityDTO to PlayerWithClubDTO as it
better describes what kind of DTO it is.

Player entity now lazily initialize its reference
to the club it belongs to. And PlayerDTO now
only has a clubId reference instead of an entire
ClubDTO

Match entity now only holds a list of player ids
instead of entire Player entities.