Update jwt and getCompetitions method
jwt now includes the clubid for the logged in user. The long-term
goal should be to check each incoming request so that the logged
in user can administer this competition.

For now, a manual check in the getCompetitions endpoint is done
so that only competitions for the club currently logged in are
returned.