Explicit types for matches
This is a PoC for having explicit types for the two
types of matches we generally consider; matches in
playoff, and matches in pool games.

This commit shows that we can have Records transformed
to explicit match types, at the same time that we can
keep the original MatchDTOs that we serve for front end.