DRAFT: Showing and manually editing seeding.
The idea is to expose two new endpoints:
- GetCurrentSeeding which return the seeding of the competition category
- ApproveSeeding which will commit the seeding to the datastore. This
method will allow user to manually commit a seeding of their liking.

We also introduce a new state called SEEDING_HAS_BEEN_COMMITTED indicating
that a seeding has been stored and committed to datastore.

Like before, if you have not committed a seeding when making a draw, we just
calculate one on the fly and store it.