TP-91 Backend implementation of allowing manual approval of seeding (#241)
This PR implements the backend functionallity required to visualize and modify
the seeding of a draw.

A side effect now is that we have to differentiate between deleting a draw and
re-opening a competition category. Deleting a draw means that the draw is
removed while the seeding is intact, just so a previously manually approved
seeding is not removed and cause double the job for a administrator. Also,
the competition category is still closed for registrations.

Re-opening a competition category means registrations can be added, this
means that we have to remove any previous draw and saved seeding.